// src/app/api/api/models/despacho-response.ts
/* tslint:disable */
/* eslint-disable */
export interface DespachoResponse {
  activo?: boolean;
  cantidad?: number;
  clienteId?: number;
  codigo?: string;
  despachoId?: number;
  destinoId?: number;
  estadoDsc?: string;
  fechaDespacho?: string;
  fechaEstimadaLlegada?: string;
  navieraId?: number;
  numeroContenedor?: string;
  observaciones?: string;
  operadorLogisticoId?: number;
  productoId?: number;
  puertoLlegadaId?: number;
  situacionId?: number;
  unidadMedida?: string;
  variedadId?: number;
  viaId?: number;
}